requirements.txt
python-gnupg
loguru
sqlalchemy
asyncpg